pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "RAG Search Implementation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.65.0",
    "huggingface-hub>=1.5.0",
    "nltk==3.9.1",
    "numpy>=2.4.2",
    "openai>=1.0.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
